renamed dumpComponent to ChatbotComponent
