+++ client/src/App.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import "./App.css";
 import io from "socket.io-client";
-import ChatBotRobot from "./dumpComponent";
+import ChatBotRobot from "./ChatbotComponent";
 
 const socket = io("http://localhost:3000", { transports: ["websocket"] });
 socket.connect(true);
